Move common functions to helpers
